+++ core-services/egov-indexer/src/main/java/org/egov/infra/indexer/consumer/PTCustomIndexMessageListener.java
@@ -4,6 +4,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.egov.infra.indexer.custom.pt.PTCustomDecorator;
 import org.egov.infra.indexer.custom.pt.PropertyRequest;
 import org.egov.infra.indexer.service.IndexerService;
+import org.egov.infra.indexer.util.DLQHandler;
 import org.egov.infra.indexer.util.IndexerUtils;
 import org.slf4j.MDC;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,6 +31,9 @@ public class PTCustomIndexMessageListener implements MessageListener<String, Str
 	@Autowired
 	private PTCustomDecorator ptCustomDecorator;
 
+	@Autowired
+	private DLQHandler dlqHandler;
+
 	@Value("${egov.indexer.pt.update.topic.name}")
 	private String ptUpdateTopic;
 
@@ -44,19 +48,25 @@ public class PTCustomIndexMessageListener implements MessageListener<String, Str
 	 */
 	public void onMessage(ConsumerRecord<String, String> data) {
 		ObjectMapper mapper = indexerUtils.getObjectMapper();
-
-
+		
+		String tenantId = null;
 		try {
 			PropertyRequest propertyRequest = mapper.readValue(data.value(), PropertyRequest.class);
 			// Adding in MDC so that tracer can add it in header
-			MDC.put(TENANTID_MDC_STRING, propertyRequest.getProperties().get(0).getTenantId() );
+			tenantId = propertyRequest.getProperties().get(0).getTenantId();
+			MDC.put(TENANTID_MDC_STRING, tenantId);
 
 			if (data.topic().equals(ptUpdateTopic))
 				propertyRequest = ptCustomDecorator.dataTransformForPTUpdate(propertyRequest);
 			propertyRequest.setProperties(ptCustomDecorator.transformData(propertyRequest.getProperties()));
 			indexerService.esIndexer(data.topic(), mapper.writeValueAsString(propertyRequest));
 		} catch (Exception e) {
-			log.error("Couldn't parse ptindex request: ", e);
+			dlqHandler.handleError(data.value(), e, "PTCustomIndexMessageListener");
+		} finally {
+			// Always clear MDC to prevent data leakage across thread reuse
+			if (tenantId != null) {
+				MDC.remove(TENANTID_MDC_STRING);
+			}
 		}
 	}
